refactor: convert to library
