+++ src/lib.rs
@@ -1,0 +1,7 @@
+pub mod core;
+pub mod network;
+mod proto;
+mod test_util;
+#[cfg(test)]
+mod tests;
+mod util;
